+++ src/chip.cc
@@ -7,6 +7,10 @@ NAN_MODULE_INIT(Chip::Init) {
   tpl->SetClassName(Nan::New("Chip").ToLocalChecked());
   tpl->InstanceTemplate()->SetInternalFieldCount(1);
 
+  Nan::SetPrototypeMethod(tpl, "getNumberOfLines", getNumberOfLines);
+  Nan::SetPrototypeMethod(tpl, "getChipName", getChipName);
+  Nan::SetPrototypeMethod(tpl, "getChipLabel", getChipLabel);
+
   constructor.Reset(Nan::GetFunction(tpl).ToLocalChecked());
   Nan::Set(target, Nan::New("Chip").ToLocalChecked(), Nan::GetFunction(tpl).ToLocalChecked());
 }
@@ -46,6 +50,45 @@ NAN_METHOD(Chip::New) {
     info.GetReturnValue().Set(Nan::NewInstance(cons, argc, argv).ToLocalChecked());
   }
   DOUT( "%s %s():%d\n", __FILE__, __FUNCTION__, __LINE__);
+}
+
+NAN_METHOD(Chip::getNumberOfLines) {
+  DOUT( "%s %s():%d\n", __FILE__, __FUNCTION__, __LINE__);
+  Chip *obj = Nan::ObjectWrap::Unwrap<Chip>(info.This());
+  if ( !obj->chip) {
+    Nan::ThrowError( "::getNumberOfLines() for chip==NULL");
+    return;
+  }
+  int ret = gpiod_chip_num_lines(obj->getNativeChip());
+  if(-1 == ret) {
+    Nan::ThrowError( "::getNumberOfLines() failed");
+  } else info.GetReturnValue().Set(ret);
+}
+
+NAN_METHOD(Chip::getChipName) {
+  DOUT( "%s %s():%d\n", __FILE__, __FUNCTION__, __LINE__);
+  Chip *obj = Nan::ObjectWrap::Unwrap<Chip>(info.This());
+  if ( !obj->chip) {
+    Nan::ThrowError( "::getChipName() for chip==NULL");
+    return;
+  }
+  const char *name = gpiod_chip_name(obj->getNativeChip());
+  if(!name) {
+    Nan::ThrowError( "::getChipName() failed");
+  } else info.GetReturnValue().Set(Nan::New<v8::String>(name).ToLocalChecked());
+}
+
+NAN_METHOD(Chip::getChipLabel) {
+  DOUT( "%s %s():%d\n", __FILE__, __FUNCTION__, __LINE__);
+  Chip *obj = Nan::ObjectWrap::Unwrap<Chip>(info.This());
+  if ( !obj->chip) {
+    Nan::ThrowError( "::getChipLabel() for chip==NULL");
+    return;
+  }
+  const char *label = gpiod_chip_label(obj->getNativeChip());
+  if(!label) {
+    Nan::ThrowError( "::getChipLabel() failed");
+  } else info.GetReturnValue().Set(Nan::New<v8::String>(label).ToLocalChecked());
 }
 
 gpiod_chip *Chip::getNativeChip() {