+++ src/chip.hh
@@ -9,7 +9,7 @@
 #if USE_PRINTF
 #define DOUT(fmt,args...) printf(fmt,##args)
 #else
-#define DOUT(fmt,args...) 
+#define DOUT(fmt,args...)
 #endif
 
 class Chip : public Nan::ObjectWrap {
@@ -20,6 +20,10 @@ class Chip : public Nan::ObjectWrap {
  private:
   explicit Chip(const char* device = "0");
   ~Chip();
+
+  static NAN_METHOD(getNumberOfLines);
+  static NAN_METHOD(getChipName);
+  static NAN_METHOD(getChipLabel);
 
   static NAN_METHOD(New);
   static Nan::Persistent<v8::Function> constructor;